Move NewChildren to it's own module
