+++ src/vivarium_gates_iv_iron/components/pregnancy.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pandas as pd
 
 from vivarium.framework.engine import Builder
@@ -8,6 +7,7 @@ from vivarium.framework.population import SimulantData
 from vivarium_gates_iv_iron.components.hemoglobin import Hemoglobin
 from vivarium_gates_iv_iron.components.mortality import MaternalMortality
 from vivarium_gates_iv_iron.components.disability import MaternalDisability
+from vivarium_gates_iv_iron.components.children import NewChildren
 
 from vivarium_gates_iv_iron.constants import models, data_keys
 from vivarium_gates_iv_iron.constants.data_values import (
@@ -312,38 +312,3 @@ class Pregnancy:
             self.new_children.empty(new_not_pregnant.index)
         )
         return new_not_pregnant
-
-
-class NewChildren:
-
-    @property
-    def name(self):
-        return 'child_status'
-
-    @property
-    def columns_created(self):
-        return ['sex_of_child', 'birth_weight']
-
-    def setup(self, builder: Builder):
-        self.randomness = builder.randomness.get_stream(self.name)
-
-    def empty(self, index: pd.Index):
-        return pd.DataFrame({
-            'sex_of_child': models.INVALID_OUTCOME,
-            'birth_weight': np.nan,
-        }, index=index)
-
-    def __call__(self, index: pd.Index):
-        sex_of_child = self.randomness.choice(
-            index,
-            choices=['Male', 'Female'],
-            additional_key='sex_of_child',
-        )
-        # TODO implement LBWSG on next line for sampling
-        draw = self.randomness.get_draw(index, additional_key='birth_weight')
-        birth_weight = 1500. * (1 + draw)
-
-        return pd.DataFrame({
-            'sex_of_child': sex_of_child,
-            'birth_weight': birth_weight,
-        }, index=index)
